+++ RELEASE-NOTES.md
@@ -1,3 +1,4 @@
 ## 0.26
 
 * #1 Support for collecting basic Amazon credentials
+* #2 Support injectable S3 Client - AmazonS3